File rename and import updates
